+++ devtools/src/components/MediaSettings.tsx
@@ -1,0 +1,66 @@
+import type React from 'react'
+import { useStore } from '../store/useStore'
+import AudioCodecMimeType from './AudioCodecMimeType'
+import AudioInputDevice from './AudioInputDevice'
+import AudioOutputDevice from './AudioOutputDevice'
+import AudioToggle from './AudioToggle'
+import CameraPermissionState from './CameraPermissionState'
+import MicrophonePermissionState from './MicrophonePermissionState'
+import RequestMediaPermissionButton from './RequestMediaPermissionButton'
+import TransceiverDirection from './TransceiverDirection'
+import VideoCodecMimeType from './VideoCodecMimeType'
+import VideoInputDevice from './VideoInputDevice'
+import VideoToggle from './VideoToggle'
+
+const MediaSettings: React.FC = () => {
+  const audioDirection = useStore((state) => state.settings.audio.direction)
+  const setAudioDirection = useStore((state) => state.setAudioDirection)
+  const videoDirection = useStore((state) => state.settings.video.direction)
+  const setVideoDirection = useStore((state) => state.setVideoDirection)
+
+  return (
+    <fieldset style={{ maxWidth: '500px' }}>
+      <legend>Media settings</legend>
+      <p>
+        <RequestMediaPermissionButton buttonText="Request Media Permission" />
+      </p>
+      <p>
+        Audio:
+        <AudioToggle />
+        <br />
+        Microphone permission state:
+        <MicrophonePermissionState />
+        <br />
+        Audio input device:
+        <AudioInputDevice />
+        <br />
+        Audio output device:
+        <AudioOutputDevice />
+        <br />
+        Direction:
+        <TransceiverDirection value={audioDirection} onChange={setAudioDirection} />
+        <br />
+        AudioCodec (MIME type):
+        <AudioCodecMimeType />
+      </p>
+      <p>
+        Video:
+        <VideoToggle />
+        <br />
+        Camera permission state:
+        <CameraPermissionState />
+        <br />
+        Video input device:
+        <VideoInputDevice />
+        <br />
+        Direction:
+        <TransceiverDirection value={videoDirection} onChange={setVideoDirection} />
+        <br />
+        Video Codec (MIME type):
+        <VideoCodecMimeType />
+      </p>
+    </fieldset>
+  )
+}
+
+export default MediaSettings