+++ devtools/vite.config.mjs
@@ -1,7 +1,9 @@
 import { resolve } from 'node:path'
+import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vite'
 
 export default defineConfig({
+  plugins: [react()],
   root: resolve(__dirname),
   dist: resolve(__dirname, 'dist'),
   resolve: {